fix(models): fixed sime relationships between models
