+++ app/api/models/gift_models.py
@@ -85,15 +85,6 @@ class Gift(Base):  # type: ignore
     payment_link = Column(
         String,
     )
-    bank_name = Column(
-        String,
-    )
-    bank_account_number = Column(
-        Integer,
-    )
-    bank_account_name = Column(
-        String,
-    )
     gift_type = Column(
         ENUM("physical", "cash", name="gift_type"), nullable=False
     )
